fix: sdk.utils.addMemoTx move to sdk.utils.sol.addMemoTx
