+++ src/utils/index.ts
@@ -1,45 +1,22 @@
-import { Connection, PublicKey, TransactionMessage, VersionedTransaction } from "@solana/web3.js";
-import { SdkError } from "../exceptions";
-import { AllbridgeCoreSdkOptions, ChainSymbol } from "../index";
+import { AllbridgeCoreSdkOptions } from "../index";
 import { NodeRpcUrlsConfig } from "../services";
-import { fetchAddressLookupTableAccountsFromTx } from "./sol";
+import { DefaultSolUtils, SolUtils } from "./sol";
 import { DefaultSrbUtils, SrbUtils } from "./srb";
 
 /**
  * Contains usefully methods
  */
 export interface Utils {
-  /**
-   * Add memo to solana's transaction
-   * @param transaction transaction to add memo
-   * @param memo memo to add (28 char max)
-   */
-  addMemoToTx(transaction: VersionedTransaction, memo: string): Promise<void>;
-
   srb: SrbUtils;
+  sol: SolUtils;
 }
 
 export class DefaultUtils implements Utils {
   srb: SrbUtils;
+  sol: SolUtils;
 
   constructor(readonly nodeRpcUrlsConfig: NodeRpcUrlsConfig, params: AllbridgeCoreSdkOptions) {
     this.srb = new DefaultSrbUtils(nodeRpcUrlsConfig, params);
-  }
-
-  async addMemoToTx(transaction: VersionedTransaction, memo: string): Promise<void> {
-    if (memo.length > 28) {
-      throw new SdkError("InvalidArgumentException memo cannot be more than 28 characters");
-    }
-    const connection = new Connection(this.nodeRpcUrlsConfig.getNodeRpcUrl(ChainSymbol.SOL), "confirmed");
-    const addressLookupTableAccounts = await fetchAddressLookupTableAccountsFromTx(transaction, connection);
-    const message = TransactionMessage.decompile(transaction.message, {
-      addressLookupTableAccounts: addressLookupTableAccounts,
-    });
-    message.instructions[message.instructions.length - 1].keys.push({
-      pubkey: new PublicKey(Buffer.from(memo)),
-      isSigner: false,
-      isWritable: false,
-    });
-    transaction.message = message.compileToV0Message(addressLookupTableAccounts);
+    this.sol = new DefaultSolUtils(nodeRpcUrlsConfig, params);
   }
 }
