+++ src/utils/sol/index.ts
@@ -1,23 +1,37 @@
-import { AddressLookupTableAccount, Connection, VersionedTransaction } from "@solana/web3.js";
-import { SdkError } from "../../exceptions";
+import { Connection, PublicKey, TransactionMessage, VersionedTransaction } from "@solana/web3.js";
+import { AllbridgeCoreSdkOptions, ChainSymbol, SdkError } from "../../index";
+import { NodeRpcUrlsConfig } from "../../services";
+import { fetchAddressLookupTableAccountsFromTx } from "./utils";
 
-export async function fetchAddressLookupTableAccountsFromTx(
-  transaction: VersionedTransaction,
-  connection: Connection
-): Promise<AddressLookupTableAccount[]> {
-  return await Promise.all(
-    transaction.message.addressTableLookups.map(async (lookup) => {
-      return new AddressLookupTableAccount({
-        key: lookup.accountKey,
-        state: AddressLookupTableAccount.deserialize(
-          await connection.getAccountInfo(lookup.accountKey).then((res) => {
-            if (!res) {
-              throw new SdkError("Cannot get AccountInfo");
-            }
-            return res.data;
-          })
-        ),
-      });
-    })
-  );
+/**
+ * Contains usefully Solana methods
+ */
+export interface SolUtils {
+  /**
+   * Add memo to solana's transaction
+   * @param transaction transaction to add memo
+   * @param memo memo to add (28 char max)
+   */
+  addMemoToTx(transaction: VersionedTransaction, memo: string): Promise<void>;
+}
+
+export class DefaultSolUtils implements SolUtils {
+  constructor(readonly nodeRpcUrlsConfig: NodeRpcUrlsConfig, readonly params: AllbridgeCoreSdkOptions) {}
+
+  async addMemoToTx(transaction: VersionedTransaction, memo: string): Promise<void> {
+    if (memo.length > 28) {
+      throw new SdkError("InvalidArgumentException memo cannot be more than 28 characters");
+    }
+    const connection = new Connection(this.nodeRpcUrlsConfig.getNodeRpcUrl(ChainSymbol.SOL), "confirmed");
+    const addressLookupTableAccounts = await fetchAddressLookupTableAccountsFromTx(transaction, connection);
+    const message = TransactionMessage.decompile(transaction.message, {
+      addressLookupTableAccounts: addressLookupTableAccounts,
+    });
+    message.instructions[message.instructions.length - 1].keys.push({
+      pubkey: new PublicKey(Buffer.from(memo)),
+      isSigner: false,
+      isWritable: false,
+    });
+    transaction.message = message.compileToV0Message(addressLookupTableAccounts);
+  }
 }
